refactor(utils): break runtime utils->metrics cycle; drop mutable defaults
Wave C structural (the one clearly-real dependency-direction bug):
- Move ScoringContext from metrics/_context to utils/_context (it is
  validation metadata, a utils concern). validate_data constructs it from
  utils now, so utils no longer imports up into metrics at runtime;
  metrics/_context re-exports it for back-compat. The residual
  utils->metrics edge is TYPE_CHECKING-only (a BaseScorer annotation).
- PolynomialTrendForecaster uses estimator=None + builds ElasticNet in
  __init__ (mirrors the FourierSeasonalityForecaster fix), so default-
  constructed instances no longer share one mutable estimator object.
